refactor(llma): extract model matching utils and fix review issues
- Extract ~270 lines of pure model matching functions into
  playgroundModelMatching.ts, breaking the circular import between
  llmPlaygroundPromptsLogic and llmPlaygroundModelLogic
- Fix isTraceLikeSelection treating empty string provider as truthy
- Remove redundant empty actions connect arrays in llmPlaygroundRunLogic
- Remove unnecessary :any type annotation in submitPrompt handler
- Inline pickProviderKeyIdForProvider single-line wrapper
